backend/requirements.txt
# ── Core ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# ── Database (Async PostgreSQL) ──
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# ── OpenAI (Cloud mode) ──
openai==1.58.1

# ── Utilities ──
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
